renaming to allow script to show up as meadowphysics in menu
