+++ data/mp.data
@@ -1,30 +1,30 @@
 return {
 -- Table: {1}
 {
-   ["trigger"]={2},
+   ["sync"]={2},
    ["speed"]={3},
-   ["scount"]={4},
-   ["prev_mode"]=0,
-   ["sync"]={5},
-   ["state"]={6},
-   ["kcount"]=0,
-   ["smin"]={7},
-   ["position"]={8},
-   ["tick"]={9},
-   ["clear"]={10},
-   ["toggle"]={11},
-   ["edit_row"]=0,
-   ["key_count"]=0,
-   ["rule_dests"]={12},
-   ["sound"]=0,
-   ["rules"]={13},
-   ["min"]={14},
-   ["rule_dest_targets"]={15},
-   ["pushed"]={16},
+   ["state"]={4},
    ["mode"]=0,
-   ["smax"]={17},
-   ["count"]={18},
-   ["max"]={19},
+   ["kcount"]=0,
+   ["rule_dests"]={5},
+   ["trigger"]={6},
+   ["prev_mode"]=0,
+   ["scount"]={7},
+   ["sound"]=0,
+   ["toggle"]={8},
+   ["rules"]={9},
+   ["max"]={10},
+   ["rule_dest_targets"]={11},
+   ["count"]={12},
+   ["tick"]={13},
+   ["pushed"]={14},
+   ["position"]={15},
+   ["smax"]={16},
+   ["key_count"]=0,
+   ["edit_row"]=0,
+   ["clear"]={17},
+   ["min"]={18},
+   ["smin"]={19},
 },
 -- Table: {2}
 {
@@ -61,6 +61,17 @@ return {
 },
 -- Table: {5}
 {
+   1,
+   2,
+   3,
+   4,
+   5,
+   6,
+   7,
+   8,
+},
+-- Table: {6}
+{
    2,
    4,
    8,
@@ -69,17 +80,6 @@ return {
    64,
    128,
    256,
-},
--- Table: {6}
-{
-   0,
-   0,
-   0,
-   0,
-   0,
-   0,
-   0,
-   0,
 },
 -- Table: {7}
 {
@@ -94,71 +94,27 @@ return {
 },
 -- Table: {8}
 {
-   4,
-   5,
-   3,
-   1,
-   2,
-   9,
-   10,
-   5,
+   0,
+   0,
+   0,
+   0,
+   0,
+   0,
+   0,
+   0,
 },
 -- Table: {9}
 {
-   0,
-   0,
-   0,
-   0,
-   0,
-   0,
-   0,
-   0,
+   2,
+   2,
+   2,
+   2,
+   2,
+   2,
+   2,
+   2,
 },
 -- Table: {10}
-{
-   0,
-   0,
-   0,
-   0,
-   0,
-   0,
-   0,
-   0,
-},
--- Table: {11}
-{
-   0,
-   0,
-   0,
-   0,
-   0,
-   0,
-   0,
-   0,
-},
--- Table: {12}
-{
-   1,
-   2,
-   3,
-   4,
-   5,
-   6,
-   7,
-   8,
-},
--- Table: {13}
-{
-   2,
-   2,
-   2,
-   2,
-   2,
-   2,
-   2,
-   2,
-},
--- Table: {14}
 {
    9,
    10,
@@ -169,7 +125,7 @@ return {
    15,
    16,
 },
--- Table: {15}
+-- Table: {11}
 {
    3,
    3,
@@ -179,6 +135,50 @@ return {
    3,
    3,
    3,
+},
+-- Table: {12}
+{
+   9,
+   10,
+   11,
+   12,
+   13,
+   14,
+   15,
+   16,
+},
+-- Table: {13}
+{
+   0,
+   0,
+   0,
+   0,
+   0,
+   0,
+   0,
+   0,
+},
+-- Table: {14}
+{
+   0,
+   0,
+   0,
+   0,
+   0,
+   0,
+   0,
+   0,
+},
+-- Table: {15}
+{
+   8,
+   8,
+   10,
+   8,
+   5,
+   4,
+   8,
+   4,
 },
 -- Table: {16}
 {
@@ -215,13 +215,13 @@ return {
 },
 -- Table: {19}
 {
-   9,
-   10,
-   11,
-   12,
-   13,
-   14,
-   15,
-   16,
+   0,
+   0,
+   0,
+   0,
+   0,
+   0,
+   0,
+   0,
 },
 }
